+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated RoadUtility to use Unity 2019.2 API
 - Added an option to change the desired height of the ramp
 - Refactor of Profiling
+- Added the ability to change default and selected gizmo color
 
 
 ## [Version: 1.9_FH] - 2019.04.22